VizUI - rename to visualizations.tsx from file
