+++ vulcan/lib/client/jsx/components/workflow/user_interactions/inputs/visualizations.tsx
